requirements.txt
dreg-client==1.0.2
humanfriendly==9.2
python-dotenv==0.19.0
simplejson==3.17.5
urwid==2.1.2
